move ssh tests to a new file
